Move stuff to util
