+++ src/main/scala/nl/vroste/zio/kinesis/client/ProducerLive.scala
@@ -45,6 +45,8 @@ private[client] final class ProducerLive[R, R1, T](
   inFlightCalls: Ref[Int],
   triggerUpdateShards: UIO[Unit]
 ) extends Producer[T] {
+  import Util.ZStreamExtensions
+
   var runloop: ZIO[Logging with Clock, Nothing, Unit] =
     // Failed records get precedence
     (ZStream
@@ -477,50 +479,4 @@ private[client] object ProducerLive {
     foldWhileM(PutRecordsAggregatedBatchForShard.empty)(_.isWithinLimits) { (batch, record: ProduceRequest) =>
       ZIO.succeed(batch.add(record))
     }.mapM(_.toProduceRequest)
-
-  implicit class ZStreamExtensions[R, E, O](stream: ZStream[R, E, O]) {
-    // ZStream's groupBy using distributedWithDynamic is not performant enough, maybe because
-    // it breaks chunks
-    final def groupByKey2[K](
-      getKey: O => K,
-      buffer: Int = 16
-    ): ZStream[R, E, (K, ZStream[Any, E, O])] =
-      ZStream.unwrapManaged {
-        type GroupQueueValues = Exit[Option[E], Chunk[O]]
-
-        for {
-          substreamsQueue     <- Queue
-                               .bounded[Exit[Option[E], (K, Queue[GroupQueueValues])]](buffer)
-                               .toManaged(_.shutdown)
-          substreamsQueuesMap <- Ref.make(Map.empty[K, Queue[GroupQueueValues]]).toManaged_
-          addToSubStream       = (key: K, values: Chunk[O]) => {
-                             for {
-                               substreams <- substreamsQueuesMap.get
-                               _          <- if (substreams.contains(key))
-                                      substreams(key).offer(Exit.succeed(values))
-                                    else
-                                      Queue
-                                        .bounded[GroupQueueValues](buffer)
-                                        .tap(_.offer(Exit.succeed(values)))
-                                        .tap(q => substreamsQueue.offer(Exit.succeed((key, q))))
-                                        .tap(q => substreamsQueuesMap.update(_ + (key -> q)))
-                                        .unit
-                             } yield ()
-                           }
-          _                   <- (stream.foreachChunk { chunk =>
-                   ZIO.foreach_(chunk.groupBy(getKey))(addToSubStream.tupled)
-                 } *> substreamsQueue.offer(Exit.fail(None))).catchSome {
-                 case e: E => substreamsQueue.offer(Exit.fail(Some(e)))
-               }.forkManaged
-        } yield ZStream.fromQueueWithShutdown(substreamsQueue).collectWhileSuccess.map {
-          case (key, substreamQueue) =>
-            val substream = ZStream
-              .fromQueueWithShutdown(substreamQueue)
-              .collectWhileSuccess
-              .flattenChunks
-            (key, substream)
-        }
-      }
-
-  }
 }
